pyproject.toml
[tool.coverage.run]
branch = true
omit = ["config.py", "config-3.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if typing.TYPE_CHECKING:",
]

[tool.mypy]
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "photonlibpy.*"
# https://github.com/PhotonVision/photonvision/issues/1210
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "--strict-markers -v --maxfail=2"
pythonpath = "."
testpaths = ["tests"]
xfail_strict = true

[tool.ruff]
target-version = "py311"
line-length = 80

[tool.ruff.lint]
preview = true
select = [
    # pycodestyle
    "E",
    # "E501",
    # pyflakes
    # "F",
    # flake8-bugbear
    # "B",
    # pyupgrade
    # "UP",
    # flake8-comprehensions
    # "C4",
    # flake8-logging-format
    # "G",
    # flake8-simplify
    # "SIM",
    # flake8-print
    # "T20",  # we'll stop making them a warning for now
    # perflint
    # "PERF",
]

[tool.uv]
dev-dependencies = [
    "hypothesis>=6.112.1",
    "mypy>=1.8.0",
    "pytest>=7.2.0",
    "pytest-integration>=0.2.3",
]
prerelease = "allow"

[project]
name = "pyreefscape"
version = "0.0.0"
description = "The Drop Bears' FRC 2025 robot code"
authors = [{ name = "The Drop Bears", email = "enquiries@thedropbears.org.au" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10,<3.13"

dependencies = [
    "numpy~=2.1",
    "phoenix6~=25.1.0",
    "robotpy==2025.1.1.1",
    "robotpy-apriltag~=2025.1.1.0",
    "robotpy-wpilib-utilities==2025.0.0",
    "photonlibpy==2025.0.0b8",
]

[tool.robotpy]
requires = [
    "numpy~=2.1",
    "phoenix6~=25.1.0",
    "robotpy==2025.1.1.1",
    "robotpy-apriltag~=2025.1.1.0",
    "robotpy-wpilib-utilities==2025.0.0",
    "photonlibpy==2025.0.0b8",
]
robotpy_version = "2025.1.1.1"
robotpy_extras = ["apriltag"]
